requirements.txt
alembic==1.8.1
certifi==2022.6.15
distlib==0.3.5
flask==2.1.3
Flask-RESTful==0.3.9
flask-swagger-ui==4.11.1
filelock==3.7.1
greenlet==1.1.2
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
pipenv==2022.7.24
platformdirs==2.5.2
pytz==2022.1
python-dateutil==2.8.2
pytest==7.1.2
six==1.16.0
SQLAlchemy==1.4.39
virtualenv==20.16.0
virtualenv-clone==0.5.7
